BACKEND/API/requirements.txt
autopep8==1.6.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
toml==0.10.2
Werkzeug==2.1.1
zipp==3.8.0